+++ src/app/cmd/add.rs
@@ -142,7 +142,7 @@ impl Add {
     fn get_content_from_stdin(&self) -> Result<Option<String>> {
         
         if !is_content_piped() {
-            return Err(anyhow!(cmd::Error::MissingPipe));
+            return Err(anyhow!(app::Error::MissingPipe));
         }
         let mut buffer = String::new();
         io::stdin().read_to_string(&mut buffer)?;