+++ src/app/mod.rs
@@ -2,8 +2,34 @@ pub mod cmd;
 pub mod config;
 pub mod creds;
 pub use config::Config;
+use std::fmt;
 
 pub struct Context {
     pub config: Option<Config>,
     pub api_token: Option<String>,
+}
+
+/// App errors are largely handled by creating anyhow::Errors
+/// But where we do need to take action depending on the concrete error,
+/// add a variant here.
+/// If wrapped in anyhow::Errors, they can be retrieved with anyhow::root_cause()
+#[derive(Debug)]
+pub enum Error {
+    MissingPipe,
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            Error::MissingPipe => {
+                write!(f, "Tried to read from stdin pipe, but nothing was piped")
+            }
+        }
+    }
+}
+
+impl std::error::Error for Error {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        None
+    }
 }